# openapi.yaml
openapi: 3.0.3
info:
  title: FRAS - Facial Recognition Attendance System
  description: API for managing attendance via facial recognition.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development
paths:
...
  /api/students:
    get:
      summary: List students
      tags: [Students]
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated list of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedStudents"
    post:
      summary: Create a new student
      tags: [Students]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/students/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get student by ID
      tags: [Students]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
    put:
      summary: Update a student
      tags: [Students]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Student not found
    delete:
      summary: Delete a student
      tags: [Students]
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Student deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Student not found
  /api/students/register-face:
    post:
      summary: Register face encoding for a student
      tags: [Students]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [student_id, image]
              properties:
                student_id:
                  type: string
                  format: uuid
                image:
                  type: string
                  format: base64
                  description: Base64-encoded face image
      responses:
        "200":
          description: Face registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  encoding_id:
                    type: string
                    format: uuid
        "400":
          description: No face detected or multiple faces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    Student:
      type: object
      properties:
        id:
          type: string
          format: uuid
        student_id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        department:
          type: string
        year:
          type: integer
        face_registered:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    StudentInput:
      type: object
      required: [student_id, name, email]
      properties:
        student_id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        department:
          type: string
        year:
          type: integer
    PaginatedStudents:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        pages:
          type: integer
...